Move relevant lifecycle plugins tests over
