+++ autodispose/src/test/java/com/uber/autodispose/AutoDisposeParallelFlowableTest.java
@@ -1,14 +1,10 @@
 package com.uber.autodispose;
 
 import io.reactivex.Flowable;
-import io.reactivex.functions.Consumer;
-import io.reactivex.functions.Predicate;
 import io.reactivex.processors.PublishProcessor;
-import io.reactivex.subjects.BehaviorSubject;
 import io.reactivex.subjects.MaybeSubject;
 import io.reactivex.subscribers.TestSubscriber;
 import java.util.List;
-import org.junit.After;
 import org.junit.Rule;
 import org.junit.Test;
 import org.reactivestreams.Subscriber;
@@ -20,10 +16,6 @@ public class AutoDisposeParallelFlowableTest {
   private static final int DEFAULT_PARALLELISM = 2;
 
   @Rule public final RxErrorsRule rule = new RxErrorsRule();
-
-  @After public void resetPlugins() {
-    AutoDisposePlugins.reset();
-  }
 
   @Test public void ifParallelism_and_subscribersCount_dontMatch_shouldFail() {
     TestSubscriber<Integer> subscriber = new TestSubscriber<>();
@@ -149,106 +141,6 @@ public class AutoDisposeParallelFlowableTest {
 
     assertThat(source.hasSubscribers()).isFalse();
     assertThat(scope.hasObservers()).isFalse();
-  }
-
-  @Test public void autoDispose_withProviderAndNoOpPlugin_withoutStarting_shouldFailSilently() {
-    AutoDisposePlugins.setOutsideLifecycleHandler(
-        new Consumer<OutsideLifecycleException>() {
-          @Override
-          public void accept(OutsideLifecycleException e) throws Exception {
-          }
-        });
-    BehaviorSubject<Integer> lifecycle = BehaviorSubject.create();
-    TestSubscriber<Integer> firstSubscriber = new TestSubscriber<>();
-    TestSubscriber<Integer> secondSubscriber = new TestSubscriber<>();
-    LifecycleScopeProvider<Integer> provider = TestUtil.makeLifecycleProvider(lifecycle);
-    PublishProcessor<Integer> source = PublishProcessor.create();
-    //noinspection unchecked
-    Subscriber<Integer>[] subscribers = new Subscriber[] {firstSubscriber, secondSubscriber};
-
-    source
-        .parallel(DEFAULT_PARALLELISM)
-        .as(AutoDispose.<Integer>autoDisposable(provider))
-        .subscribe(subscribers);
-
-    assertThat(source.hasSubscribers()).isFalse();
-    assertThat(lifecycle.hasObservers()).isFalse();
-
-    firstSubscriber.assertNoValues();
-    firstSubscriber.assertNoErrors();
-    secondSubscriber.assertNoValues();
-    secondSubscriber.assertNoErrors();
-  }
-
-  @Test public void autoDispose_withProviderAndNoOpPlugin_afterEnding_shouldFailSilently() {
-    AutoDisposePlugins.setOutsideLifecycleHandler(
-        new Consumer<OutsideLifecycleException>() {
-          @Override
-          public void accept(OutsideLifecycleException e) {
-            // Noop
-          }
-        });
-    BehaviorSubject<Integer> lifecycle = BehaviorSubject.createDefault(0);
-    lifecycle.onNext(1);
-    lifecycle.onNext(2);
-    lifecycle.onNext(3);
-    TestSubscriber<Integer> firstSubscriber = new TestSubscriber<>();
-    TestSubscriber<Integer> secondSubscriber = new TestSubscriber<>();
-    LifecycleScopeProvider<Integer> provider = TestUtil.makeLifecycleProvider(lifecycle);
-    PublishProcessor<Integer> source = PublishProcessor.create();
-    //noinspection unchecked
-    Subscriber<Integer>[] subscribers = new Subscriber[] {firstSubscriber, secondSubscriber};
-
-    source
-        .parallel(DEFAULT_PARALLELISM)
-        .as(AutoDispose.<Integer>autoDisposable(provider))
-        .subscribe(subscribers);
-
-    assertThat(source.hasSubscribers()).isFalse();
-    assertThat(lifecycle.hasObservers()).isFalse();
-    firstSubscriber.assertNoValues();
-    firstSubscriber.assertNoErrors();
-  }
-
-  @Test public void autoDispose_withProviderAndPlugin_withoutStarting_shouldFailWithExp() {
-    AutoDisposePlugins.setOutsideLifecycleHandler(
-        new Consumer<OutsideLifecycleException>() {
-          @Override
-          public void accept(OutsideLifecycleException e) {
-            throw new IllegalStateException(e);
-          }
-        });
-    BehaviorSubject<Integer> lifecycle = BehaviorSubject.create();
-    TestSubscriber<Integer> firstSubscriber = new TestSubscriber<>();
-    TestSubscriber<Integer> secondSubscriber = new TestSubscriber<>();
-    LifecycleScopeProvider<Integer> provider = TestUtil.makeLifecycleProvider(lifecycle);
-    PublishProcessor<Integer> source = PublishProcessor.create();
-    //noinspection unchecked
-    Subscriber<Integer>[] subscribers = new Subscriber[] {firstSubscriber, secondSubscriber};
-
-    source
-        .parallel(DEFAULT_PARALLELISM)
-        .as(AutoDispose.<Integer>autoDisposable(provider))
-        .subscribe(subscribers);
-
-    firstSubscriber.assertNoValues();
-    firstSubscriber.assertError(
-        new Predicate<Throwable>() {
-          @Override
-          public boolean test(Throwable throwable) {
-            return throwable instanceof IllegalStateException
-                && throwable.getCause() instanceof OutsideLifecycleException;
-          }
-        });
-    secondSubscriber.assertNoValues();
-    secondSubscriber.assertError(
-        new Predicate<Throwable>() {
-          @Override
-          public boolean test(Throwable throwable) throws Exception {
-            return throwable instanceof IllegalStateException
-                && throwable.getCause() instanceof OutsideLifecycleException;
-          }
-        });
   }
 
   @Test public void autoDispose_withScopeProviderCompleted_shouldNotReportDoubleSubscriptions() {
