+++ lifecycle/autodispose-lifecycle/src/test/java/com/uber/autodispose/lifecycle/LifecycleScopeProviderSubscriberTest.java
@@ -19,11 +19,14 @@ package com.uber.autodispose.lifecycle;
 import com.uber.autodispose.AutoDispose;
 import com.uber.autodispose.RxErrorsRule;
 import io.reactivex.Flowable;
+import io.reactivex.functions.Consumer;
+import io.reactivex.functions.Predicate;
 import io.reactivex.processors.PublishProcessor;
 import io.reactivex.subjects.BehaviorSubject;
 import io.reactivex.subscribers.TestSubscriber;
 import java.util.List;
 import org.junit.After;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 
@@ -34,7 +37,7 @@ public class LifecycleScopeProviderSubscriberTest {
 
   @Rule public RxErrorsRule rule = new RxErrorsRule();
 
-  @After public void resetPlugins() {
+  @Before @After public void resetPlugins() {
     AutoDisposeLifecyclePlugins.reset();
   }
 
@@ -96,6 +99,69 @@ public class LifecycleScopeProviderSubscriberTest {
     List<Throwable> errors = o.errors();
     assertThat(errors).hasSize(1);
     assertThat(errors.get(0)).isInstanceOf(LifecycleEndedException.class);
+  }
+
+  @Test public void autoDispose_withProviderAndNoOpPlugin_withoutStarting_shouldFailSilently() {
+    AutoDisposeLifecyclePlugins.setOutsideLifecycleHandler(new Consumer<OutsideLifecycleException>() {
+      @Override public void accept(OutsideLifecycleException e) { }
+    });
+    BehaviorSubject<Integer> lifecycle = BehaviorSubject.create();
+    LifecycleScopeProvider<Integer> provider = makeLifecycleProvider(lifecycle);
+    PublishProcessor<Integer> source = PublishProcessor.create();
+    TestSubscriber<Integer> o = source
+        .as(AutoDispose.<Integer>autoDisposable(provider))
+        .test();
+
+    assertThat(source.hasSubscribers()).isFalse();
+    assertThat(lifecycle.hasObservers()).isFalse();
+    o.assertNoValues();
+    o.assertNoErrors();
+  }
+
+  @Test public void autoDispose_withProviderAndNoOpPlugin_afterEnding_shouldFailSilently() {
+    AutoDisposeLifecyclePlugins.setOutsideLifecycleHandler(new Consumer<OutsideLifecycleException>() {
+      @Override public void accept(OutsideLifecycleException e) {
+        // Noop
+      }
+    });
+    BehaviorSubject<Integer> lifecycle = BehaviorSubject.createDefault(0);
+    lifecycle.onNext(1);
+    lifecycle.onNext(2);
+    lifecycle.onNext(3);
+    LifecycleScopeProvider<Integer> provider = makeLifecycleProvider(lifecycle);
+    PublishProcessor<Integer> source = PublishProcessor.create();
+    TestSubscriber<Integer> o = source
+        .as(AutoDispose.<Integer>autoDisposable(provider))
+        .test();
+
+    assertThat(source.hasSubscribers()).isFalse();
+    assertThat(lifecycle.hasObservers()).isFalse();
+    o.assertNoValues();
+    o.assertNoErrors();
+  }
+
+  @Test public void autoDispose_withProviderAndPlugin_withoutStarting_shouldFailWithExp() {
+    AutoDisposeLifecyclePlugins.setOutsideLifecycleHandler(new Consumer<OutsideLifecycleException>() {
+      @Override public void accept(OutsideLifecycleException e) {
+        // Wrap in an IllegalStateException so we can verify this is the exception we see on the
+        // other side
+        throw new IllegalStateException(e);
+      }
+    });
+    BehaviorSubject<Integer> lifecycle = BehaviorSubject.create();
+    LifecycleScopeProvider<Integer> provider = makeLifecycleProvider(lifecycle);
+    PublishProcessor<Integer> source = PublishProcessor.create();
+    TestSubscriber<Integer> o = source
+        .as(AutoDispose.<Integer>autoDisposable(provider))
+        .test();
+
+    o.assertNoValues();
+    o.assertError(new Predicate<Throwable>() {
+      @Override public boolean test(Throwable throwable) {
+        return throwable instanceof IllegalStateException
+            && throwable.getCause() instanceof OutsideLifecycleException;
+      }
+    });
   }
 
 }
